requirements.txt
pandas
numpy
scipy
scikit-learn
python-dotenv
langchain
langchain-openai
langfuse>=3,<4
ulid-py
beautifulsoup4
lxml
pyyaml
pytest
